requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic
scikit-learn
nump
pandas
scipy
